Add authorization to posts/comments delete/editing

+++ src/main/java/dev/gyuray/forum/service/CommentService.java
@@ -2,6 +2,7 @@ package dev.gyuray.forum.service;
 
 import dev.gyuray.forum.domain.Comment;
 import dev.gyuray.forum.domain.Post;
+import dev.gyuray.forum.domain.Role;
 import dev.gyuray.forum.domain.User;
 import dev.gyuray.forum.repository.comment.CommentForm;
 import dev.gyuray.forum.repository.comment.CommentListDTO;
@@ -15,6 +16,7 @@ import org.springframework.transaction.annotation.Transactional;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @Transactional
@@ -27,14 +29,14 @@ public class CommentService {
     private final UserService userService;
 
     @Transactional
-    public Long addComment(CommentForm commentForm) {
+    public Long addComment(CommentForm commentForm, User user) {
         Comment comment = new Comment(commentForm.getContent());
 
         Long postId = commentForm.getPostId();
         Post foundPost = postService.findPostById(postId);
         comment.addToPost(foundPost);
 
-        Long userId = commentForm.getUserId();
+        Long userId = user.getId();
         User foundUser = userService.findUser(userId);
         comment.addToUser(foundUser);
 
@@ -60,13 +62,27 @@ public class CommentService {
     }
 
     @Transactional
-    public void updateComment(CommentUpdateDTO commentUpdateDTO) {
-        commentRepository.update(commentUpdateDTO);
+    public void updateComment(Long commentId, CommentUpdateDTO commentUpdateDTO, User user) {
+        Comment foundComment = findCommentById(commentId);
+
+        if (foundComment.getUser().getId() == user.getId()) {
+            foundComment.updateComment(commentUpdateDTO);
+        } else {
+            log.info("foundComment.getUser().getLoginId() = {}", foundComment.getUser().getLoginId());
+            log.info("user = {}", user.getLoginId());
+            throw new IllegalStateException("댓글을 수정할 권한이 없습니다.");
+        }
     }
 
     @Transactional
-    public void deleteComment(Long commentId) {
-        commentRepository.delete(commentId);
+    public void deleteComment(Long commentId, User user) {
+        Comment foundComment = findCommentById(commentId);
+
+        if (foundComment.getUser().getId() == user.getId() || user.getRole() == Role.ADMIN) {
+            commentRepository.delete(commentId);
+        } else {
+            throw new IllegalStateException("댓글을 삭제할 권한이 없습니다.");
+        }
     }
 
 }
